pyproject.toml
[project]
name = "tor-monitor-pro"
version = "1.0.0"
description = "Professional Tor relay monitoring with multi-relay support, alerts, and enterprise features"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Tor Monitor Pro Team"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "stem>=1.8.0",
    "rich>=13.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "prometheus-client>=0.19.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "cryptography>=41.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "itsdangerous>=2.1.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=23.0.0", "mypy>=1.0.0", "ruff>=0.1.0"]
postgres = ["psycopg2-binary>=2.9.0"]
ldap = ["python-ldap>=3.4.0"]

[project.scripts]
tor-monitor-pro = "tor_monitor_pro.cli:main"
tor-monitor-pro-check = "tor_monitor_pro.cli:check_setup"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"